mkmetalink.py: move code to proper package for reuse
The script remains as a simple wrapper.

QubesOS/qubes-issues#1254
