+++ .gitignore
@@ -1,1 +1,2 @@
 current-release
+qubesmirror/**/*.py[co]
